Correct call of os.system
This patch fix the correct call of os.system before you were calling something
like:
    os.system("./main a b c d")

which was wrong in so many ways

    os.system("./main 20 10 10 2")

That is correct and in order to cal it right you need to create the string
first correctly. That part is fixed now

Then you need to have an image before everything run , otherwise it will create
an error and stop there . That was the initial problem

Then yoru button is not right , you need a handler. right now is.system is
being executed all the time , not when you click the button

Then you need to reload the image, after it is generated

With all those changes you are good to go :)

Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>